Add Alpha Construct

+++ Helpers/ProjectileHelpers.cs
@@ -14,28 +14,44 @@ public class ProjectileHelpers { }
 /// </summary>
 public abstract class AuraProjectile : ModProjectile
 {
+	public Vector2 OwnerCenter;
 	public Color AuraDrawColor { private get; set; } = Color.White;
 	public float AuraDrawColorMultiplier { private get; set; } = 1f;
 
 	public Player Owner => Main.player[Projectile.owner];
 
 	public abstract bool ShouldKillProjectile();
+
+	public virtual Vector2 GetCenter() => Owner.MountedCenter;
+
 	public abstract void SafeSetDefaults();
 
 	public override void SetDefaults() {
 		Projectile.penetrate = -1;
 		Projectile.tileCollide = false;
 		Projectile.hide = true;
+		Projectile.alpha = 255;
 		SafeSetDefaults();
 	}
 
 	public override void AI() {
 		Projectile.velocity = Vector2.Zero;
-		Projectile.Center = Owner.MountedCenter;
+		Projectile.Center = GetCenter();
 		Projectile.timeLeft = 2;
+		Projectile.rotation += 0.01f;
 
 		if (ShouldKillProjectile()) {
-			Projectile.Kill();
+			Projectile.alpha += 12;
+			if (Projectile.alpha >= 255) {
+				Projectile.Kill();
+			}
+
+			return;
+		}
+
+		Projectile.alpha -= 12;
+		if (Projectile.alpha < 0) {
+			Projectile.alpha = 0;
 		}
 	}
 
@@ -45,8 +61,9 @@ public abstract class AuraProjectile : ModProjectile
 		Vector2 drawPosition = Projectile.Center.ToScreenCoordinates();
 		Rectangle sourceRect = new(0, 0, texture.Width, texture.Height);
 		Vector2 origin = sourceRect.Size() / 2f;
-		Color drawColor = AuraDrawColor * GeneralUtils.GetBrightness(Projectile.Center) * AuraDrawColorMultiplier;
-		Main.spriteBatch.Draw(texture, drawPosition, sourceRect, drawColor, 0f, origin, 1f, SpriteEffects.None, 0f);
+		Color drawColor = AuraDrawColor * GeneralUtils.GetBrightness(Projectile.Center) * AuraDrawColorMultiplier * Projectile.Opacity;
+		float scale = MathHelper.Lerp(0.5f, 1f, Projectile.Opacity);
+		Main.spriteBatch.Draw(texture, drawPosition, sourceRect, drawColor, Projectile.rotation, origin, scale, SpriteEffects.None, 0f);
 
 		return false;
 	}
